# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service provide functionality for storing and managing privat telegram channels with subscription based payments for acessing content.",
        "title": "Telegram Subs API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/account/{acc_id}/user": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "updates telegram user conected to this account. Admin, can update email for any user. Either of one of the fields must be provided",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update TgUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "acc_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "account new email",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateAccountUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/user": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "creates user with given users telegram_id and username. Only administrator and bot can create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.createUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.createUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get user list. Only administrator and bot can get list of accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "List Users",
                "parameters": [
                    {
                        "description": "user list request filter",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.FindUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Tguser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/{usr_id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "returns user object by given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "usr_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.getUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "deletes user object by given id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "usr_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/user/{usr_id}/password": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "updateds users username and/or status, provided by id. Admin and bot can update any user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "usr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user new status or username",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.updateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.FindUserRequest": {
            "type": "object",
            "properties": {
                "page_settings": {
                    "description": "Username *struct {\n\tLike string `json:\"LIKE\"`\n\tEq   string `json:\"EQ\"`\n} `json:\"username\"`\nEmail *struct {\n\tLike string `json:\"LIKE\"`\n\tEq   string `json:\"EQ\"`\n} `json:\"email\"`",
                    "type": "object",
                    "properties": {
                        "page_number": {
                            "type": "integer"
                        },
                        "page_size": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "domain.Tguser": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "http.createUserRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.createUserResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.getUserResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "http.updateAccountUserRequest": {
            "type": "object",
            "properties": {
                "telegram_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "http.updateUserRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "http.updateUserResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}